build: refactor db service and entity service
