+++ apps/api/src/app/auth/auth.service.ts
@@ -1,11 +1,12 @@
 import * as bcrypt from 'bcrypt';
 import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
+import { Observable } from 'rxjs';
+import { UpdateWriteOpResult } from 'mongodb';
+import { map } from 'rxjs/operators';
 import { UserDto } from '../users/dto/user.dto';
 import { LoginDto } from './dto/login.dto';
 import { UsersService } from '../users/users.service';
-import { Observable } from 'rxjs';
-import { UpdateWriteOpResult } from 'mongodb';
 
 @Injectable()
 export class AuthService {
@@ -66,6 +67,7 @@ export class AuthService {
         { $set: { lastLogIn: new Date() } },
         { upsert: true }
       )
+      .pipe(map((result) => result.value))
       .toPromise();
 
     if (!user) {
